refactor: 게임 진행 로직 controller 로 재배치
